add filters and validation to set/create helpers
These helpers are going to be used by other routes in our site, so it's
valuable to be able to validate the data. If we want to bypass
validation and filtering, we can work with the backend directly (don't).

We also expose the validation as its own helper just in case.