merge db schema into configuration
This greatly reduces the amount of configuration a user is required to
specify, and helps prevent the user repeating themselves.

There is likely a lot of work that needs to be added to make sure the
schema is auto-detected correctly, but that will be persued via
developing some sample applications.