fix login form requiring authentication
If we put authentication on our entire site, and the site has a default
route that slurps up strings like `login` or `logout`, this would cause
the login form to be behind authentication. So, unauthenticated users
would not be allowed to see the login form to turn them into
authenticated users.

Example app that demonstrates problem before this commit fixes it:

    use Mojolicious::Lite;
    my $base = any( '' );
    plugin 'Yancy', backend => 'test://localhost';
    app->yancy->plugin( 'Auth::Basic', { route => $base } );
    my $auth = app->yancy->auth->route;
    $auth->get( '/:anything', sub {
        shift->render( data => 'Authed' );
    } );

In this app, the `/:anything` route would be matched when the user tried
to visit `/login`. Instead of seeing the login form, they would see an
error message saying they were not logged in.

This happened because the Mojolicious router goes through routes in
order of creation, depth-first, to find the first one that matches.
Since the `$base` route matches anything, its children are searched
next. Then it finds `/:anything` which matches anything, too.

The fix is to create `/login` and `/logout` before the
authentication-required route. That way `/login` is hit first. In
general, it's better for routes to be created from more-specific to
less-specific.