add exception for when table lacks primary key
In the event a table is missing a primary key, creating new items works
but fetching those items will never work. So we're unable to return the
item that we just created, which makes the API fail data validation.

Now we throw a proper exception instead of just returning nothing.

Refs #10