fix url sending user to page that doesn't exist
If a bunch of items got deleted, the user could find themselves on
a page that doesn't exist. Then they'd be forced to hit the `<` button
a bunch of times to get to real pages.

Now if we have somehow got to a page that doesn't exist, we just go to
the first page instead.