move main api controller to allow for others
In the future we're going to ship some controllers that users can use to
simplify their apps, and it'd be nicer if we could take the good name of
`Yancy`.

Since we're going to be providing controllers and explainations for
using controllers, we need to be more specific about what the
`controller_class` is: A controller for the API used by the Yancy editor
(and not likely a thing that users will want to override).