add id field to user table in tests
It is more real for a table to always have an ID and to instead have
a unique constraint for a username. The ID that Yancy cares about could
be any unique ID, not just the primary key of a file, and now we have
tests for it.

It's also a useful case to test for, and revealed a few bugs in our
backends:

1. The DBIx::Class backend couldn't handle any other key than the
   primary key
2. MySQL was returning the ID it created, not necessarily the ID that
   Yancy cares about.