allow removing collections added via read_schema
We can't just hide collections that we don't want to work with: They
still need to be configured correctly for the schema to be valid and the
API to work.

Now, the new `x-ignore` configuration causes the collection to be
removed entirely from the API, which removes it from the editor.