use any unique constraint during read_schema
This solves the `mojo_migrations` problem, provided all of the supported
backends add a unique constraint there. Not sure why it's not a primary
key (perhaps to keep the index out of memory).

So, the `x-ignore` feature isn't necessary to solve this problem, but
it's still useful enough to keep around.