remove `null` values from items in API
The OpenAPI plugin performs validation of the output from the API, which
is good. However, this leads to a problem with optional fields: If they
are not explicitly marked as allowing `null` (by setting `type:
[ <type>, "null" ]`), an error will be returned instead.

We added documentation to correct this, but it's far easier to just fix
this automatically for the user by removing any field whose value is
`undef` (`null` in JavaScript terms).

We already do this in the editor application, which is how these values
are allowed to be written to the database in the first place. Now we do
the same for data being returned by the API to the editor application.

Fixes #12