fix warnings generated by JSON::Validator
Yancy uses formats to both define how a field should be validated, and
to define how a field should be displayed. If the format is not one of
JSON::Validator's known formats, it (rightly) warns us that it doesn't
know what to do. This is annoying, so now we make sure that all
supported formats are defined, even if we can't actually validate them.

Refs #13