fix getting new item in collection with x-id-field
After we create a new item, we get the ID of the item, and then we have
to fetch the item's data. To do this, we take the URL in the spec and
fill in any fields (denoted in `{field_name}`).

However, the only field that the code was filling in was named "id".
This is not true if the collection renames its ID field with
`x-id-field`. This would lead to a URL like
`/yancy/api/users/undefined`, which is (hopefully) not a valid ID.

Thanks @ViktorNacht for reporting this issue!

Fixes #17