try to show the x-id-field in the default list
When we aren't given a list of columns we have to try to make one up. If
we then also don't have a field named "id", it is quite possible that
our table will have no columns in it at all. Now, we use the x-id-field
if specified to make sure that we have at least one useful column in the
table.

Thanks @ViktorNacht for reporting this problem

Refs #17