fix unauthorized api response becoming 404
The new version of Mojolicious::Plugin::OpenAPI turns this 401 response
we were generating into a 404: The `openapi` argument doesn't enable the
OpenAPI handler, so Mojolicious looks for a template instead. Since no
`template` was set, it looks for `.html.ep` as a template, which doesn't
exist and ends up as a "not found" error.

This is because the route that will respond with the OpenAPI spec is not
itself handled by the OpenAPI handler. We might end up needing to detect
if we're in an OpenAPI route before trying to use the OpenAPI handler.

This, along with the release of a new plugin version that fixes the
default handling, should fix issue #18