allow partial updates via `set` helper
Partial updates assume that the required fields are already in the
database. Since we don't want to slow things down by forcing a fetch
from the database to complete the item before validation, we modify the
validation routine to only validate the partial data.

I'm worried about this change making it easier to bypass validation, but
it's the only way I can think of to allow forms with partial fields.