disallow other props when doing partial validation
We're restricting the validation schema to only the given fields. To
avoid using this to bypass validation, we must make absolutely sure that
no other fields are passed-in.

This has some repercussions for existing `additionalProperties` in the
schema: If the property isn't specifically-defined, it won't validate.
We can fix that by specifically defining all properties passed-in for
partial validation by using the `additionalProperties` schema.