test validation of date fields
This is a step towards fixing non-required date/date-time fields in the
Yancy controller `set` action: The action should remove these fields
before trying to validate the item.

The problem is that forms include every named field in the result, even
ones that weren't set. For some fields, that's okay: We can coerce
a good-enough default value. For other fields, though, we need to do
what the user intended and remove the field from the data.