sync test schemas with test data
All of the data that the tests use needs to be expressible in all of the
backend schemas that we support. This is tested with the
share/run_backend_tests.pl script, and by Travis. Over the last few
weeks, we've made a bunch of changes to the test data but have not fixed
the test schemas to match. This commit fixes all of that.