normalize boolean values to 1 and 0
The different databases have very different ways of dealing with
booleans:

* SQLite has no separate boolean type, but will store booleans as
  integers (0 and 1)
* MySQL translates `BOOLEAN` to `TINYINT(1)`, so is also an integer (0
  and 1)
* Postgres supports a boolean type, and lots of values will get
  normalized to `t` (true) and `f` (false) in the database. Moreover,
  DBD::Pg translates the `t` and `f` to `1` and `0` for Perl.

JSON::Validator (boolean coersions on) allows a bunch of values for
booleans as well. It's these values that all backends should support.

Now the SQLite and MySQL backends normalize booleans to 0 and 1 before
saving for ease. The DBIC backend also does this to support SQLite and
MySQL as underlying databases.