fix detection of unique keys in sqlite
SQL is not case-sensitive, and `read_schema` for SQLite requires parsing
the SQL used to create the table. So, a column can be `UNIQUE` or
`unique` or `UnIqUe` and we need to be able to detect it as such.

This prevents Yancy failing to start because of Mojo::SQLite::Migrations
`mojo_migrations` table.