fix string 0/1 failing boolean validation
The JSON::Validator coersions try to support `0` and `1` as booleans,
but it doesn't support `"0"`and `"1"`, which makes it impossible for
forms to submit booleans other than `"true"` and `"false"`. Now we
pre-process booleans to normalize them always into `"true"` and
`"false"` for validation, which then gets normalized to `1` and `0` when
stored in the database.