add better error message when missing id field
Yancy needs a unique ID field to be able to work with the data. If it
doesn't find one, it tries to assume there's an `id` field. Except if
that property doesn't exist, the OpenAPI plugin throws an error like

    Possibly a typo in schema? Could not find
    "/definitions/<table>/properties/id" in ""
    (#/definitions/<table>/properties/id)

This error says nothing about what the problem really is or potential
remedies.

Now there's a sanity check that tries to ensure there's a useful ID
field for every configured collection and throws an exception if there
is not.

Thanks @mario-minati for the issue report!

Fixes #35