rename "collection" to "schema"
The idea here is that "schema" is easier to say, but also in the future
where we put "schema" can be either the _name_ of a schema or an actual
schema (useful for Form plugins and Controllers).

This commit leaves all the tests largely as they were, triggering all
the deprecation warnings as a test. The next commit will remove all the
deprecation triggers in the tests, as well as make the deprecation
warnings slightly less annoying...

Refs #25