deprecate Yancy::Controller::Yancy::API
This means merging all the functionality of the API controller into the
standard controller, Yancy::Controller::Yancy. This reduces the amount
of code in Yancy, but also ensures that the standard controller is
robust enough to handle common use-cases like supporting a single-page
application.

In the future, there may be an external Yancy::Plugin::OpenAPI to add
back the OpenAPI response validation. Eventually, the OpenAPI plugin
will no longer be required at all by Yancy...