fix yancy editor for new yancy controller
Changing over to the Yancy::Controller::Yancy controller for the editor
caused some bugs in the editor itself: The `create` API route returns
the new item, not just the ID of the new item.

Additionally, the editor application was not setting the correct
`Content-Type` header for the request, so we try a little harder to get
JSON in the controller (which may come back to bite me, but we'll try it
and see...)

We also have now started building Selenium-based tests for the JS
application. These will be filled in more as regressions need testing,
new features get added, and proper documentation is added.

Thanks Markus@irc.freenode.net#mojo for the report.