add custom menu items to the yancy editor
Now users can add their own views to the editor. This is very
preliminary, we probably need to add VueX and Vue-Router to make this
a sane application, but this gets the idea out the door for now.

I also automated screenshot creation via the selenium test so we can
keep updating the screenshots as Yancy itself changes.