allow additional backends in controller, editor
Adding custom backends to the editor required adding them to the
controller as well. Adding custom backends to the controller required
allowing a request-wide backend so that all the backend helpers (list,
get, set, create, delete) would work correctly. This may prove to be
a terrible decision, as it means that a controller can only ever work
with one backend at a time, but we'll see how it goes...