fix required password not passing validation
If the password field is required, we must always set it. This forces
the user to always "change" their password every time they send the form
to update the rest of a user object. So, to pass validation, we fill in
password fields with dummy data to satisfy the requirement.