fix controller to respond to any supported format
If we do not provide a fallback `any` to `respond_to`, any other format
get a blank `204 No Content` response. But, if we do provide an `any`
handler to `respond_to`, the `format` gets blanked and an HTML template is
rendered.

So, we need to save the format before `respond_to` and then restore it
inside the `any` handler. This does the right thing.