fix no registration fields causing error (#69)
When there are no "required" fields declared in the schema and no
explicit "register_fields" passed-in to the Password plugin, we should
fall back to an empty arrayref. Previously, this was dying with
a bizarre error later when it tried to use the register fields: `Can't
use an undefined value as an ARRAY reference`.

Thanks @ViktorNacht for reporting this issue!

Fixes #69