fix date/time validation of empty strings
@chy-causer reported an issue with validating date/time fields from custom
forms: Form text fields that exist are always sent, even if there's no
value inside (the empty string). This was failing JSON::Validator
validation since the empty string does not match an ISO8601 date/time
field. This was fixed by changing '' to `null` before validation (which
is what the editor does, but before submission). There may be other
things that need this same behavior.

To test this, I needed to add date/time fields to the test schemas. It
didn't make sense to have these _not_ default to the current date/time,
so I did that. Turns out that's a lot harder than I thought: Database
differences are fun, but also supporting the "now" special value was
interesting. It can be used as a default, but not as input from the user
(for now, until we can figure out a sane way to make the JSON schema
support both 'now' and '2020-01-01T00:00:00', maybe regex...)