make login_form work on pluggable auth
The Yancy::Plugin::Auth module's login_form helper behaved differently
than all the others: It would render a complete page. This makes it
impossible to include in to another page. Now it only shows the form and
there is a separate 'login page' to display a page.

This is a breaking change:

* The template `yancy/auth/login.html.ep` has been moved to
  `yancy/auth/login_page.html.ep`
* The template `yancy/auth/password/login.html.ep` has been moved to
  `yancy/auth/password/login_page.html.ep` (to match above)