tighten up auth workflow
This reduces the number of page loads a user has to go through to
authenticate: If the user is unauthenticated and goes to an unauthorized
page, the login form is displayed. If the login form is displayed on any
page other than a direct request to the login page, they are returned to
where they were when auth is complete. Logging out takes the user back
to where they were before they logged out (so, if necessary, they can
log in again).