fix problems with selenium not finding the toast
This just started happening on my machine: Sometimes Selenium will wait
for the toast to appear, but then when it tries to click it throws an
error about it not being in the DOM anymore. This might be something to
do with Vue and caching, but I don't care enough to figure that out yet.