fix dbic backend not finding foreign keys
DBIx::Class result classes do not need the same name as the table,
though that's what we use in Yancy's tests to make the DBIx::Class
schema look exactly the same as all the other schemas. When the result
class has a different name, we want to use that name to avoid breaking
the encapsulation of the ORM (the underlying table is not useful
information, since it may change).

Thanks @mario-minati for help debugging this! Fixes #112