allow all joins to reference schema
It makes no sense to reference a local table property in the "join": We
aren't joining the property, we're joining the schema the property
refers to.

This also adds the potential future ability to have a key that isn't the
primary key of the foreign table by having a "." in the "x-foreign-key"
value.