ignoring column defaults of db which are function calls issue #124
currently read_schema returns the default values of db in verbatim
even if they are function calls. This causes attempts to insert
the function call body as defaults in columns instead of the
the value, which does not makes sense. The current fix returns undef
for such column defaults. Additionally as a measure of caution it
does not do the above in case the default has been handled in
fix_default mapping.