support for complex joins on list method via json encoding.
support for specifying complex joins via query parameter is being
introduced. The json is encoded as per uri encoding scheme and is
decoded into a perl reference. The same reference is passed as the
'join' parameter in DBIC. This allows the same familier
syntax to be adopted for specifying the joins.

How it works:
* The data structure in join is passed to a function _get_unique_relationships
to get all the unique relations as keys of a hash. for each of such relation
the properties are determined and are stored in another hash having keys
in the format 'relation_name.key_name' , the value being the properties of
the column. these properties help in using the appropriate operators at sql
level. Note that it is also possible to filter the results based on any of
the relationship columns by specifying as relation.columnname=somevalue in
query parameters.