fix missing and unwanted routes
We really only want endpoints. We don't want 'under' to show up in the
list of pages, because it isn't. We also don't want websockets, as
there's no _content_ there (now that I think about it, we probably
really only want 'GET' routes...)

This also adds a 'hidden' stash value to make a route not appear in the
list. I'm using this for the editor and API routes so they can't,
themselves, be edited.